+++ src/main/java/wegrus/clubwebsite/entity/post/BoardCategory.java
@@ -1,4 +1,4 @@
-package wegrus.clubwebsite.entity.board;
+package wegrus.clubwebsite.entity.post;
 
 public enum BoardCategory {
     GROUP, STUDY, BOARD
