+++ src/main/java/wegrus/clubwebsite/entity/post/BoardType.java
@@ -1,4 +1,4 @@
-package wegrus.clubwebsite.entity.board;
+package wegrus.clubwebsite.entity.post;
 
 public enum BoardType {
     IXPLOIT, IGDC, ALGROUS, WEBGROUS, PERSONAL, FREE, INFO, PROJECT, HOBBY, SUGGEST
