+++ src/main/java/wegrus/clubwebsite/repository/CommentLikeRepository.java
@@ -4,8 +4,8 @@ import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.data.jpa.repository.Modifying;
 import org.springframework.data.jpa.repository.Query;
 import org.springframework.data.repository.query.Param;
-import wegrus.clubwebsite.entity.board.CommentLike;
-import wegrus.clubwebsite.entity.board.Reply;
+import wegrus.clubwebsite.entity.post.CommentLike;
+import wegrus.clubwebsite.entity.post.Reply;
 import wegrus.clubwebsite.entity.member.Member;
 
 import java.util.Optional;
