+++ src/main/java/wegrus/clubwebsite/repository/PostLikeRepository.java
@@ -4,8 +4,8 @@ import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.data.jpa.repository.Modifying;
 import org.springframework.data.jpa.repository.Query;
 import org.springframework.data.repository.query.Param;
-import wegrus.clubwebsite.entity.board.Post;
-import wegrus.clubwebsite.entity.board.PostLike;
+import wegrus.clubwebsite.entity.post.Post;
+import wegrus.clubwebsite.entity.post.PostLike;
 import wegrus.clubwebsite.entity.member.Member;
 
 import java.util.Optional;
