+++ src/main/java/wegrus/clubwebsite/repository/PostRepository.java
@@ -1,7 +1,7 @@
 package wegrus.clubwebsite.repository;
 
 import org.springframework.data.jpa.repository.JpaRepository;
-import wegrus.clubwebsite.entity.board.Post;
+import wegrus.clubwebsite.entity.post.Post;
 
 public interface PostRepository extends JpaRepository<Post, Long> {
