+++ src/main/java/wegrus/clubwebsite/repository/ReplyRepository.java
@@ -4,8 +4,8 @@ import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.data.jpa.repository.Modifying;
 import org.springframework.data.jpa.repository.Query;
 import org.springframework.data.repository.query.Param;
-import wegrus.clubwebsite.entity.board.Post;
-import wegrus.clubwebsite.entity.board.Reply;
+import wegrus.clubwebsite.entity.post.Post;
+import wegrus.clubwebsite.entity.post.Reply;
 
 public interface ReplyRepository extends JpaRepository<Reply, Long> {
