+++ src/main/java/wegrus/clubwebsite/service/PostService.java
@@ -4,11 +4,11 @@ import lombok.RequiredArgsConstructor;
 import org.springframework.security.core.context.SecurityContextHolder;
 import org.springframework.stereotype.Service;
 import org.springframework.transaction.annotation.Transactional;
-import wegrus.clubwebsite.dto.board.PostCreateRequest;
-import wegrus.clubwebsite.dto.board.PostUpdateRequest;
-import wegrus.clubwebsite.entity.board.Post;
-import wegrus.clubwebsite.entity.board.PostState;
-import wegrus.clubwebsite.entity.board.PostLike;
+import wegrus.clubwebsite.dto.post.PostCreateRequest;
+import wegrus.clubwebsite.dto.post.PostUpdateRequest;
+import wegrus.clubwebsite.entity.post.Post;
+import wegrus.clubwebsite.entity.post.PostState;
+import wegrus.clubwebsite.entity.post.PostLike;
 import wegrus.clubwebsite.entity.member.Member;
 import wegrus.clubwebsite.exception.*;
 import wegrus.clubwebsite.repository.*;
