+++ src/main/java/wegrus/clubwebsite/service/ReplyService.java
@@ -4,11 +4,11 @@ import lombok.RequiredArgsConstructor;
 import org.springframework.security.core.context.SecurityContextHolder;
 import org.springframework.stereotype.Service;
 import org.springframework.transaction.annotation.Transactional;
-import wegrus.clubwebsite.dto.board.ReplyCreateRequest;
-import wegrus.clubwebsite.entity.board.Post;
-import wegrus.clubwebsite.entity.board.CommentLike;
-import wegrus.clubwebsite.entity.board.Reply;
-import wegrus.clubwebsite.entity.board.ReplyState;
+import wegrus.clubwebsite.dto.post.ReplyCreateRequest;
+import wegrus.clubwebsite.entity.post.Post;
+import wegrus.clubwebsite.entity.post.CommentLike;
+import wegrus.clubwebsite.entity.post.Reply;
+import wegrus.clubwebsite.entity.post.ReplyState;
 import wegrus.clubwebsite.entity.member.Member;
 import wegrus.clubwebsite.exception.*;
 import wegrus.clubwebsite.repository.PostRepository;
